+++ packages/snap/src/index.ts
@@ -17,9 +17,6 @@ import { SnapKvStore } from "./snapdb";
 import * as JSON from "bigint-json-serialization";
 import { loadNocturneConfigBuiltin } from "@nocturne-xyz/config";
 import { panel, text, heading } from "@metamask/snaps-ui";
-import { createLogger } from "winston";
-import BrowserConsoleTransport from "winston-transport-browserconsole";
-import { GetNotesOpts } from "@nocturne-xyz/sdk/dist/src/NocturneDB";
 
 // Local
 // const RPC_URL = "http://127.0.0.1:8545/";
@@ -42,14 +39,6 @@ const BUNDLER_URL = "https://bundler.nocturnelabs.xyz";
 const SUBGRAPH_API_URL =
   "https://api.goldsky.com/api/public/project_cldkt6zd6wci33swq4jkh6x2w/subgraphs/nocturne/0.1.18-alpha/gn";
 const config = loadNocturneConfigBuiltin("sepolia");
-
-// logger
-const LOG_LEVEL = "debug";
-const logger = createLogger({
-  exceptionHandlers: [new BrowserConsoleTransport()],
-  rejectionHandlers: [new BrowserConsoleTransport()],
-  transports: [new BrowserConsoleTransport({ level: LOG_LEVEL })],
-});
 
 const Fr = BabyJubJub.ScalarField;
 
@@ -94,7 +83,7 @@ export const onRpcRequest: OnRpcRequestHandler = async ({
   console.log("Snap Nocturne Canonical Address: ", signer.canonicalAddress());
 
   const merkleProver = await SparseMerkleProver.loadFromKV(kvStore);
-  const syncAdapter = new SubgraphSDKSyncAdapter(SUBGRAPH_API_URL, logger);
+  const syncAdapter = new SubgraphSDKSyncAdapter(SUBGRAPH_API_URL);
   const sdk = new NocturneWalletSDK(
     signer,
     provider,
@@ -103,8 +92,7 @@ export const onRpcRequest: OnRpcRequestHandler = async ({
     nocturneDB,
     syncAdapter,
     new MockEthToTokenConverter(),
-    new BundlerOpTracker(BUNDLER_URL),
-    logger
+    new BundlerOpTracker(BUNDLER_URL)
   );
 
   console.log("Switching on method: ", request.method);